src/main/java/com/shalimov/onlineshop/dao/parser/ProductParser.java
package com.shalimov.onlineshop.dao.parser;

import com.shalimov.onlineshop.entity.Product;

import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.List;

public class ProductParser {

    public List<Product> getProducts(ResultSet resultSet) throws SQLException {
        List<Product> products = new ArrayList<>();
        while (resultSet.next()) {
            Product product = new Product();
            product.setId(resultSet.getInt("id"));
            product.setTitle(resultSet.getString("title"));
            product.setPrice(resultSet.getDouble("price"));
            product.setDescription(resultSet.getString("description"));
            products.add(product);
        }
        return products;
    }
}
